Change how it calculates map distance from player
Previously the smaller difference of the two coordinates was used as the distance, the possible tiles creating a plus shape as wide and high as the map. With the changes it would instead be a 45° rotated square with sides of 2*getMinTreasureDistance().
Alternatively, changing Math.min to Math.max would make it a simple square.